+++ cmd/local/conf/defaults_test.go
@@ -42,6 +42,7 @@ func TestSetViperDefaults(t *testing.T) {
 		key      string
 		expected interface{}
 	}{
+		{conf.KeyDisableRESTAPI, false},
 		{conf.KeyCollectAccelerationLog, false},
 		{conf.KeyCollectAccessLog, false},
 		{conf.KeyCollectAuditLog, false},